+++ what_to_wear_outdoors/cli.py
@@ -1,50 +1,27 @@
-
 """Console script for what_to_wear_outdoors."""
 import datetime as dt
 import sys
 import re
 import click
-from weather_observation import Weather, Forecast
-from clothing_options_ml import Running
 import logging
 import textwrap
-from train_model import train
+
+from what_to_wear_outdoors.outfit_predictors import RunningOutfitPredictor, RunningOutfitTranslator
+
+if __name__ == '__main__' or __package__ == '':
+    from weather_observation import Weather, Forecast
+else:
+    from .weather_observation import Weather, Forecast
 
 # TODO: Manage Command Line Arguments
 #  (-u for update model (with Excel), (-f to ask for forecast) (-d for default config) (no flags for walk-through)
-Colors = {'Title': 'blue', 'Description': 'cyan', 'Prompt': 'yellow', 'Error': 'red', 'Output': 'green'}
-
-days_of_the_week = ['mon', 'tue', 'wed', 'thu', 'fri', 'sat', 'sun']
-NOW = dt.datetime.now()
-TITLE = f'#{"-"*18} What to Wear Outdoors {"-"*18}#'
-
-''' Abstract the prompt for activity date/time in case we have to ask more than once to get it right
-'''
-
-
-def prompt_for_date_time():
-    activity_date = click.prompt(
-        click.style("\n\nWhat day will you be going outside? (Sunday, Monday, etc. "
-                    "also today, tomorrow work as well.) ", fg=Colors['Prompt']),
-        default='today', value_proc=figure_out_date)
-    if (type(activity_date) == str):
-        activity_date = figure_out_date(activity_date)
-    logging.debug(f'Activity date (determined): {activity_date}')
-
-    activity_hour = click.prompt(click.style("What hour of day will you be going outside? "
-                                             "(24 hour format or HH AM/PM)?", fg=Colors['Prompt']),
-                                 default=NOW.hour + 1, value_proc=parse_time)
-    return activity_date.combine(activity_date.date(), dt.time(hour=activity_hour))
-
-    # @click.command()
-
 
 '''
 Here's what I want the interface to look like - 
     wtw --athlete --activity outdoor activity for advice on clothing default='run' options 'run, road, mtb' [cmd]
     [cmd] update - updates the model
             --filename using the filename specified here default='../data/what i wore running.xlsx'
-          test - short-hand to execute without the prompts
+          demo-mode - short-hand to execute without the prompts
             --activity outdoor activity for advice on clothing default='run' options 'run, road, mtb'
             --duration=duration
             --wind_speed
@@ -57,46 +34,14 @@ Here's what I want the interface to look like -
             --loc - location, either zip or [city,state] default=72712   
 '''
 
-# @click.option('--date', default=dt.datetime.now().date(), help='date that you will be going outside', prompt=True)
-# @click.option('--day', default=dt.datetime.now().weekday())
-# @click.option('--hour', default=dt.datetime.now().hour, help='hour that you will be going outside', prompt=True,
-#               type=click.IntRange(0-23))
-# @click.option('--activity', type=click.Choice(['run','bike']), prompt=True,
-#               help='the type of activity you will be doing outside')
-# @click.option('--zip', help='the zipcode of the location for your activity',
-#             prompt='The zipcode for the location where you''ll be going out?', type=click.IntRange(0-99999))
-# def main(date, hour, zip, activity,day):
-def main():
-    click.clear()
-    click.secho(f'#{"-"*(len(TITLE)-2)}#', fg=Colors['Title'])
-    click.secho(TITLE, fg=Colors['Title'])
-    click.secho(f'#{"-"*(len(TITLE)-2)}#', fg=Colors['Title'])
-    [click.secho(line, fg=Colors['Description'])
-     for line in textwrap.wrap('This application is meant to provide an idea'
-                               ' of the weather conditions be during your next '
-                               'outdoor activity (within the next week).\n'
-                               'It will provide some suggestions on what to wear during that activity based '
-                               'on temperature, wind speed chance of rain etc.\n\n', 60)]
+Colors = {'Title': 'blue', 'Description': 'cyan', 'Prompt': 'yellow', 'Error': 'red', 'Output': 'green',
+          'Alternate_Output': 'cyan'}
 
-    click.pause()
-    click.clear()
-    forecast_dt = prompt_for_date_time()
-    while (forecast_dt <= NOW):
-        click.secho('\nSorry you must choose a date/time that is after the top of the next hour.\n', fg=Colors['Error'])
-        forecast_dt = prompt_for_date_time()
+days_of_the_week = ['mon', 'tue', 'wed', 'thu', 'fri', 'sat', 'sun']
+valid_dow_value = ['today', 'tomorrow'] + days_of_the_week
 
-    activity_location = click.prompt(
-        click.style("Where are you going outside (zip or city,state)? ",
-                    fg=Colors['Prompt']),
-        default='Fayetteville, AR')
-
-    w = Weather()
-    logging.debug(f'Forecast (calculated): {forecast_dt}')
-    fct = w.get_forecast(forecast_dt, activity_location)
-    print(fct)
-    running = Running()
-    print('\n')
-    [click.secho(li, fg=Colors['Output']) for li in textwrap.wrap(running.build_reply(fct))]
+NOW = dt.datetime.now()
+TITLE = f'#{"-" * 18} What to Wear Outdoors {"-" * 18}#'
 
 
 def parse_time(time_string):
@@ -112,8 +57,10 @@ def parse_time(time_string):
 
 
 def figure_out_date(weekday):
-    if (weekday == 'today'): return NOW
-    if (weekday == 'tomorrow'): return NOW + dt.timedelta(days=1)
+    if weekday == 'today':
+        return NOW
+    if weekday == 'tomorrow':
+        return NOW + dt.timedelta(days=1)
     dow_target = days_of_the_week.index(weekday.lower()[:3])
     dow_today = dt.datetime.weekday(NOW)
     days_ahead = dow_target - dow_today
@@ -121,9 +68,130 @@ def figure_out_date(weekday):
         days_ahead = + 7
     return dt.datetime.today() + dt.timedelta(days=days_ahead)
 
+@click.group()
+def cli():
+    pass
 
+
+@click.command('train_models')
+# @click.argument('datapath', type=click.File(), default='data/what i wore running.xlsx')
 def train_models():
-    train()
+    #TODO: Support using more than one model type to train with
+    rop = RunningOutfitPredictor()
+    rop.rebuild_models()
+
+
+@click.command('demo')
+@click.option('--duration', default=30, help='number of minutes you will be out')
+@click.option('--windspeed', default=0, help='windspeed (mph) at time you will be out')
+@click.option('--temp', default=75, help='forecast temperature for activity')
+@click.option('--humidity', default=50.0, help='forecast humidity')
+def demo_mode(duration, windspeed, temp, humidity):
+    '''
+    This mode is used to specfic the forecast rather than to look up a forecast and depend on the results.
+
+    :type windspeed: float
+    :param duration: number of minutes that the activity will last
+    :param windspeed: wind speed (mph) at the time of the activity
+    :param temp: forecasted (feels like) temperature (F)
+    :param humidity: percent humidity
+    :return: None
+    '''
+    hum_pct = humidity if humidity < 1 else humidity / 100
+    click.secho(f'\nRecommendation for forecast\n\tTemp:{temp}'
+                f'\n\tWind:{windspeed}\n\tHumidity:{hum_pct*100}%\n\tActivity Duration:{duration}\n')
+
+    rop = RunningOutfitPredictor()
+    conditions = dict(feels_like=temp, wind_speed=windspeed, pct_humidity=humidity,
+                      duration=duration, is_light=True)
+
+    outfit = rop.predict_outfit(**conditions)
+    rot = RunningOutfitTranslator()
+    reply = rot.build_reply(outfit, conditions)
+    [click.secho(li, fg=Colors['Output']) for li in textwrap.wrap(reply)]
+
+
+@click.command('auto')
+#@click.command('--dow')
+#@click.command('--hour')
+#@click.command('--location')
+def auto_mode(dow, hour, location):
+    print('got into auto_mode')
+    input()
+    w = Weather()
+    forecast_dt = parse_time(hour)
+    logging.debug(f'Forecast (calculated): {forecast_dt}')
+    fct = w.get_forecast(forecast_dt, location)
+    print(fct)
+    rop = RunningOutfitPredictor()
+    conditions = dict(feels_like=fct.feels_like_f, wind_speed=fct.wind_speed, pct_humidity=fct.humidity,
+                      duration=45, is_light=True)
+    rop = rop.predict_outfit(**conditions)
+    rot = RunningOutfitTranslator()
+    reply = rot.build_reply(rop.outfit_, **conditions)
+    print('\n')
+    [click.secho(li, fg=Colors['Output']) for li in textwrap.wrap(reply)]
+
+
+
+@click.command('main')
+def main():
+    click.clear()
+    click.secho(f'#{"-" * (len(TITLE) - 2)}#', fg=Colors['Title'])
+    click.secho(TITLE, fg=Colors['Title'])
+    click.secho(f'#{"-" * (len(TITLE) - 2)}#', fg=Colors['Title'])
+    [click.secho(line, fg=Colors['Description'])
+     for line in textwrap.wrap('This application is meant to provide an idea'
+                               ' of the weather conditions be during your next '
+                               'outdoor activity (within the next week).\n'
+                               'It will provide some suggestions on what to wear during that activity based '
+                               'on temperature, wind speed chance of rain etc.\n\n', 60)]
+
+    click.pause()
+    click.clear()
+    forecast_dt = prompt_for_date_time()
+    while forecast_dt <= NOW:
+        click.secho('\nSorry you must choose a date/time that is after the top of the next hour.\n', fg=Colors['Error'])
+        forecast_dt = prompt_for_date_time()
+
+    activity_location = click.prompt(
+        click.style("Where are you going outside (zip or city,state)? ",
+                    fg=Colors['Prompt']),
+        default='Fayetteville, AR')
+
+    activity_duration = click.prompt(click.style("How long will you be out?", fg=Colors['Prompt']), default=45)
+
+    w = Weather()
+    logging.debug(f'Forecast date (calculated): {forecast_dt}')
+    fct = w.get_forecast(forecast_dt, activity_location)
+    print(fct)
+    rop = RunningOutfitPredictor()
+    conditions = dict(feels_like=fct.feels_like_f, wind_speed=fct.wind_speed, pct_humidity=fct.humidity,
+                      duration=activity_duration, is_light=True)
+
+    rop.predict_outfit(**conditions)
+    rot = RunningOutfitTranslator()
+    reply = rot.build_reply(rop.outfit_
+                            , conditions)
+    click.secho(f'\n#{"-" * 18}ML OUTPUT{"-" * 18}#',fg=Colors['Output'])
+    [click.secho(li, fg=Colors['Output']) for li in textwrap.wrap(reply)]
+
+
+def prompt_for_date_time():
+    '''  Abstract the prompt for activity date/time in case we have to ask more than once to get it right
+    :return: a representation of the date and time in UTC format
+    '''
+    activity_date = click.prompt(
+        click.style("\n\nWhat day will you be going outside?", fg=Colors['Prompt']),
+        default='today', type=click.Choice(valid_dow_value))
+    activity_date = figure_out_date(activity_date)
+    logging.debug(f'Activity date (determined): {activity_date}')
+
+    activity_hour = click.prompt(click.style("What hour of day will you be going outside? "
+                                             "(24 hour format or HH AM/PM)?", fg=Colors['Prompt']),
+                                 default=NOW.hour + 1, value_proc=parse_time)
+    return activity_date.combine(activity_date.date(), dt.time(hour=activity_hour))
+
 
 if __name__ == "__main__":
     sys.exit(main())  # pragma: no cover